+++ ApiControllers/GlobalUsings.cs
@@ -1,2 +1,3 @@
 ﻿global using DataAccess.Models;
 global using DataAccess.Readers;
+global using DataAccess.Writers;
