+++ ApiControllers/Program.cs
@@ -1,3 +1,4 @@
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
@@ -7,6 +8,7 @@ builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 builder.Services.AddSingleton<IReadPerfume,PerfumeReader>();
+builder.Services.AddSingleton<IWritePerfume, PerfumeWriter>();
 
 
 var app = builder.Build();
